studioghiblimovielist/requirements.txt
django==3.0
django-redis==4.12.1
requests==2.24.0
celery==5.0.0
dj-database-url==0.5.0
psycopg2==2.8.6
djangorestframework==3.12.1